refactored

+++ src/main/java/calculator/IOExpression.java
@@ -3,14 +3,13 @@ package calculator;
 import java.util.Scanner;
 
 public class IOExpression {
-
-    public static String InputExpression(){
+    public static String InputExpression() {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Put expression");
         return scanner.nextLine().replaceAll(" ", "");
     }
-
-    public static void OutputExpression(int result){
+    
+    public static void OutputExpression(double result) {
         System.out.println(result);
     }
 }
